refactor user interactive flow from register
